refactor(cli): move command helpers from _* to internal/*
Rename packages/cli/src/commands/_envelope-overlays.ts,
_resolve-helpers.ts, and _shared.ts to commands/internal/<name>.ts
and update all import sites. Underscore-prefixed siblings are easy
to miss in directory listings and inconsistent with the rest of the
codebase; an internal/ subdir is more conventional.
