+++ packages/cli/src/commands/auth/login.ts
@@ -7,7 +7,7 @@ import {
 } from '../../core/auth'
 import { writeCredentials } from '../../core/config-store'
 import { emitInfo, emitSuccess, type OutputOptions } from '../../core/output'
-import { type GlobalFlags, resolveContext } from '../_shared'
+import { type GlobalFlags, resolveContext } from '../internal/shared'
 
 export async function run(flags: GlobalFlags, out: OutputOptions) {
   const ctx = await resolveContext(flags, out)
