+++ packages/cli/src/commands/auth/logout.ts
@@ -1,6 +1,6 @@
 import { deleteCredentials } from '../../core/config-store'
 import { emitSuccess, type OutputOptions } from '../../core/output'
-import { type GlobalFlags } from '../_shared'
+import { type GlobalFlags } from '../internal/shared'
 
 export async function run(_flags: GlobalFlags, out: OutputOptions) {
   await deleteCredentials()
