+++ packages/cli/src/commands/auth/whoami.ts
@@ -1,7 +1,7 @@
 import { readCredentials } from '../../core/config-store'
 import { MxsError } from '../../core/errors'
 import { emitSuccess, type OutputOptions } from '../../core/output'
-import { type GlobalFlags, resolveContext } from '../_shared'
+import { type GlobalFlags, resolveContext } from '../internal/shared'
 
 export async function run(flags: GlobalFlags, out: OutputOptions) {
   const ctx = await resolveContext(flags, out)
