+++ packages/cli/src/commands/config/edit.ts
@@ -1,7 +1,7 @@
 import { runEditorRoundTrip } from '../../core/editor'
 import { MxsError } from '../../core/errors'
 import { emitInfo, emitSuccess, type OutputOptions } from '../../core/output'
-import { buildApiClient, type GlobalFlags, resolveContext } from '../_shared'
+import { buildApiClient, type GlobalFlags, resolveContext } from '../internal/shared'
 
 export async function run(flags: GlobalFlags, out: OutputOptions) {
   const ctx = await resolveContext(flags, out)
