+++ packages/cli/src/commands/config/get.ts
@@ -1,5 +1,5 @@
 import { emitSuccess, type OutputOptions } from '../../core/output'
-import { buildApiClient, type GlobalFlags, resolveContext } from '../_shared'
+import { buildApiClient, type GlobalFlags, resolveContext } from '../internal/shared'
 
 export async function run(key: string, flags: GlobalFlags, out: OutputOptions) {
   const ctx = await resolveContext(flags, out)
