+++ packages/cli/src/commands/config/set.ts
@@ -1,6 +1,6 @@
 import { MxsError } from '../../core/errors'
 import { emitSuccess, type OutputOptions } from '../../core/output'
-import { buildApiClient, type GlobalFlags, resolveContext } from '../_shared'
+import { buildApiClient, type GlobalFlags, resolveContext } from '../internal/shared'
 
 export interface ConfigSetFlags {
   type?: 'json' | 'string' | 'number' | 'bool'
