+++ packages/cli/src/commands/page/create.ts
@@ -1,6 +1,6 @@
 import { emitSuccess, type OutputOptions } from '../../core/output'
 import { buildPagePayload, type PageFlagInputs } from '../../core/payload'
-import { buildApiClient, type GlobalFlags, resolveContext } from '../_shared'
+import { buildApiClient, type GlobalFlags, resolveContext } from '../internal/shared'
 
 export async function run(
   _id: unknown,
