+++ packages/cli/src/commands/page/update.ts
@@ -3,7 +3,7 @@ import { MxsError } from '../../core/errors'
 import { emitSuccess, type OutputOptions } from '../../core/output'
 import { buildPagePayload, type PageFlagInputs } from '../../core/payload'
 import { isSnowflakeId } from '../../core/resolve'
-import { buildApiClient, type GlobalFlags, resolveContext } from '../_shared'
+import { buildApiClient, type GlobalFlags, resolveContext } from '../internal/shared'
 
 export async function run(
   slugOrId: string,
