+++ packages/cli/src/commands/post/delete.ts
@@ -1,6 +1,6 @@
 import { MxsError } from '../../core/errors'
 import { emitSuccess, type OutputOptions } from '../../core/output'
-import { buildApiClient, type GlobalFlags, resolveContext } from '../_shared'
+import { buildApiClient, type GlobalFlags, resolveContext } from '../internal/shared'
 import { resolvePostId } from './resolve'
 
 export async function run(
