+++ packages/cli/src/commands/post/update.ts
@@ -1,7 +1,7 @@
 import { emitSuccess, type OutputOptions } from '../../core/output'
 import { buildPostPayload, type PostFlagInputs } from '../../core/payload'
-import { buildResolver, resolveCategoryRefs } from '../_resolve-helpers'
-import { buildApiClient, type GlobalFlags, resolveContext } from '../_shared'
+import { buildResolver, resolveCategoryRefs } from '../internal/resolve-helpers'
+import { buildApiClient, type GlobalFlags, resolveContext } from '../internal/shared'
 import { resolvePostId } from './resolve'
 
 export async function run(
