+++ packages/cli/src/commands/topic/create.ts
@@ -1,5 +1,5 @@
 import { emitSuccess, type OutputOptions } from '../../core/output'
-import { buildApiClient, type GlobalFlags, resolveContext } from '../_shared'
+import { buildApiClient, type GlobalFlags, resolveContext } from '../internal/shared'
 
 export interface TopicCreateFlags {
   name: string
